Fix a signedness compiler warning.
svn path=/branches/BRANCH_6-3/; revision=5328

+++ transact.c
@@ -1337,7 +1337,7 @@ int readheaders(int sock,
     *cp++ = '\0';
     n = stuffline(ctl, buf);
 
-    if (n == strlen(buf))
+    if ((size_t)n == strlen(buf))
 	return PS_SUCCESS;
     else
 	return PS_SOCKET;
